web: overload json.add to accept SensorType directly
tuck the static_casts away into json::add function

+++ src/web/handler/devicerequesthandler.cpp
@@ -1211,7 +1211,7 @@ namespace dss {
       boost::shared_ptr<DeviceSensor_t> sensor = pDevice->getSensor(id);
       JSONWriter json;
       json.add("sensorIndex", id);
-      json.add("sensorType", static_cast<int>(sensor->m_sensorType));
+      json.add("sensorType", sensor->m_sensorType);
       return json.successJSON();
 
     } else if (_request.getMethod() == "getSensorEventTableEntry") {
